Moving EvaluationTable component along with other solution-related components.
